fix: custom OpenAI service cannot directly inherit from OpenAI #473
